Try to fix all existing issues with benchmarks
Only considered benchmark definitions that are participating, according to category-structure.yml
Let the category-structure precisely reflect the categories each verifier participates in.
Before, we had cases like 'brick participates in ReachSafety, but opts out of ReachSafety.
In reality, it only runs on ReachSafety-Floats.'

This is now changed to 'brick participates in no meta-category, but opts in to ReachSafety-Floats.'