rule change: exclude NoDataRace from violation witness validation
There is currently no agreed-upon format for violation witnesses wrt.
the no-data-race property [1], and there do not exist any validators.

Thus it was agreed in the community meeting on Nov 7th 2023 [2], that
the NoDataRace category is excluded from violation witness validation.
The intention is that this change is only for SV-COMP 2024.
By SV-COMP 2025, there should be validator support for this property.

[1] https://gitlab.com/sosy-lab/benchmarking/sv-benchmarks/-/issues/1126
[2] https://docs.google.com/document/d/1pyh9Ln_a8AZGr1CSmzPee6MkiHK0fBhb7Kd4ZzOyOEk