Merge small software-systems categories in benchmark definitions
As proposed and decided in https://gitlab.com/sosy-lab/benchmarking/sv-benchmarks/-/issues/1384\#note_1632028102.
This was previously implemented as a change in the set files in
https://gitlab.com/sosy-lab/benchmarking/sv-benchmarks/-/merge_requests/1475,
but was reverted in https://gitlab.com/sosy-lab/benchmarking/sv-benchmarks/-/commit/6a611de315da26fd1d4c9bd888a22df38f572fce
because it produced issues with Test-Comp.
Some more explanations are available in MR
https://gitlab.com/sosy-lab/benchmarking/sv-benchmarks/-/merge_requests/1486
which was also reverted.