Allow verifiers to add XML extensions in 'extensions/{tool_name}.ext'
The content is added at the end of the XML benchmark definition.
We only add an extension for CPAchecker's column definitions.
This column definition exists in other XML files, but I checked those and find them invalid.
Based on the tools' SV-COMP 2024 logs, these tools do not produce the output required for these columns.
So I assume that this was just a copy&paste issue.